backend/requirements.txt
# HopeLink Backend Dependencies
# ==============================

# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0

# Database
supabase>=2.3.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Google Calendar Integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Environment Variables
python-dotenv>=1.0.0

# OpenAI (AI 챗봇용 - 선택적)
openai>=1.10.0

# Utilities
httpx>=0.26.0
